deepface-service/requirements.txt
flask
flask-cors
deepface
tf-keras
opencv-python-headless
numpy
Pillow
